refactor(wp-source): refactor to new actions
