+++ packages/wp-source/src/__tests__/mocks/populate.ts
@@ -1,4 +1,4 @@
-import WpSource from "../../../../";
+import WpSource from "../../..";
 
 type Populate = WpSource["libraries"]["source"]["populate"];
