--- packages/wp-source/src/source/utils/index.ts
@@ -1,1 +1,0 @@
-export { default as normalizePath } from "./normalize-path";
