+++ packages/wp-source/src/state.ts
@@ -1,8 +1,9 @@
-import WpSource from "../../";
+import WpSource from "..";
 import normalizePath from "./utils/normalize-path";
 
 const state: WpSource["state"]["source"] = {
-  data: state => pathOrLink => state.source.dataMap[normalizePath(pathOrLink)],
+  data: ({ state }) => pathOrLink =>
+    state.source.dataMap[normalizePath(pathOrLink)],
   dataMap: {},
   category: {},
   tag: {},
