+++ packages/wp-source/src/utils/normalize-path.ts
